+++ src/features/player-registration/FormStep1/helper/defaultValues.ts
@@ -1,0 +1,6 @@
+export const defaultValues = {
+  name: '',
+  username: '',
+  pronoun: '',
+  dateOfBirth: '',
+}
