+++ src/features/player-registration/FormStep1/helper/getLocalStorage.ts
@@ -1,0 +1,4 @@
+export function getLocalStorage<T>(key: string, defaultValue: T): T {
+  const savedData = localStorage.getItem(key)
+  return savedData ? JSON.parse(savedData) : defaultValue
+}
