+++ src/features/player-registration/FormStep1/stories/FormStep1.stories.tsx
