+++ src/helpers/badWordsFilter/portugueseBadWords.ts
